Remove realtime monitoring components
